requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
frozenlist==1.4.1
h11==0.14.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
PySocks==1.7.1
pytesseract==0.3.10
python-dotenv==1.0.1
schedule==1.2.1
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4
